requirements.txt
# Core
python-dotenv==1.0.0

# Data collection
feedparser==6.0.10              # RSS feeds
requests==2.31.0                # HTTP requests (for SociaVault, web scraping)
beautifulsoup4==4.12.2          # Web scraping (future phases)

# AI/Analysis
openai==1.12.0                  # OpenAI GPT-4o (Content Analyzer Agent)
anthropic==0.18.1               # Claude SDK (for AI agents: Source Discovery, Quality Monitor, Internal Data Collection)
boto3==1.34.0                   # AWS SDK (if using Bedrock for agents)

# Email delivery
sib-api-v3-sdk==7.6.0           # Brevo email

# Utilities
python-dateutil==2.8.2

# PDF Export
markdown-pdf==1.2.0             # Markdown to PDF conversion
weasyprint==60.2                # HTML to PDF (more features, needs system deps)
